+++ meeting_service/app/api/routers/task_router.py
@@ -1,7 +1,7 @@
 from typing import List
 
-from app import db
 from app.crud import task_crud
+from app.db import db
 from app.schemas import task_schemas
 from fastapi import APIRouter, Depends, HTTPException
 from sqlalchemy.ext.asyncio import AsyncSession
